requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
groq>=0.4.0
openai>=1.3.7
pandas>=2.2.0
pdfplumber>=0.10.3
pypdf2>=3.0.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
email-validator>=2.1.0
python-dateutil>=2.8.2
aiofiles>=23.2.1
sqlalchemy>=2.0.23

